lark-adapter/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lark-adapter"
version = "0.1.0"
description = "Lark/Feishu open-platform event adapter that bridges to AgentFlow's review state machine."
requires-python = ">=3.11"
authors = [{ name = "AgentFlow" }]
license = { text = "Proprietary" }
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.6",
    "httpx>=0.27",
    "pycryptodome>=3.20",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.setuptools.packages.find]
include = ["lark_adapter*"]
exclude = ["tests*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
